fix: make price as double type

+++ src/main/java/org/skypro/skyshop/product/SimpleProduct.java
@@ -10,7 +10,7 @@ public class SimpleProduct extends Product {
     }
 
     @Override
-    public int getPrice() {
+    public double getPrice() {
         return this.price;
     }
 }
